fix all 1

+++ app/Http/Controllers/CalendarController.php
@@ -20,7 +20,12 @@ class CalendarController extends Controller
     public function index()
     {
         $calendars = Calendar::all();
-
+        // $calendars[] =
+        // foreach ($calendars[0] as $calendar) {
+        //     foreach ($calendar->tickets as $ticket) {
+        //         $calendars['seat'][] = $ticket->seats->count();
+        //     }
+        // }
         return view('admin.calendar.list', ['calendars' => $calendars]);
     }
 
@@ -59,7 +64,10 @@ class CalendarController extends Controller
 
         $dateShow = $request->date_show;
         $roomID = $request->room_id;
-        $roomName = Room::where('id', $roomID)->get();
+
+        $roomName = Room::find($roomID);
+        $theaterName = $roomName->theater()->get();
+
         $filmID = $request->film_id;
 
         $flag = true;
@@ -88,7 +96,7 @@ class CalendarController extends Controller
             }
             return redirect('admin/calendars')->with('success', 'Thêm thành công');
         } else {
-            return redirect('admin/calendars/create')->with('error', 'Lỗi, phim đã có lịch chiếu ngày '.$dateShow .' ở ' . $roomName[0]->name);
+            return redirect('admin/calendars/create')->withErrors('Lỗi, phim đã có lịch chiếu ngày '.$dateShow .' ở ' . $roomName->name . ' rạp ' . $theaterName[0]->name);
         }
     }
 
@@ -108,8 +116,10 @@ class CalendarController extends Controller
     public function edit($id)
     {
         $calendar = Calendar::find($id);
+        $theaters = Theater::all();
         $rooms = Room::where('theater_id', $calendar->room->theater->id)->get();
-        return view('admin.calendar.edit', ['calendar' => $calendar, 'rooms' => $rooms]);
+
+        return view('admin.calendar.edit', compact('calendar', 'theaters', 'rooms'));
     }
 
     /**
@@ -122,47 +132,66 @@ class CalendarController extends Controller
     public function update(CalendarRequest $request, $id)
     {
         $calendar = Calendar::find($id);
-        $calendar->room_id = $request->room_id;
-        $calendar->date_show = $request->date_show;
-        $calendar->save();
+        $calDatas = Calendar::all();
 
-        // update times
-        $arrayType = $request->types;
-        $arrayTypeID = $request->types_id;
-        $arrayTime = $request->time_shows;
+        $dateShow = ($request->date_show) ? $request->date_show : null;
+        $roomID = $request->room_id;
+        $filmID = $request->film_id;
 
-        for ($i=0; $i < count($arrayTypeID); $i++) {
-            $time = CalendarTime::find($arrayTypeID[$i]);
-            $time->type_ticket = $arrayType[$i];
-            $time->time_show = $arrayTime[$i];
-            $time->calendar_id = $calendar->id;
-            $time->save();
+        $roomName = Room::find($roomID);
+        $theaterName = $roomName->theater()->get();
+
+        $flag = true;
+        foreach ($calDatas as $cal) {
+            if ($roomID == $cal->room_id && $filmID == $cal->film_id && $dateShow == $cal->date_show) {
+                $flag = false;
+                break;
+            }
         }
 
-        // add time if user choose
-        if ($request->has('types_add')) {
-            $arrTypeAdd = $request->types_add;
-            $arrTime = $request->time_shows_add;
+        if ($flag) {
+            if (isset($dateShow)) {
+                $calendar->date_show = $dateShow;
+            }
+            $calendar->room_id = $request->room_id;
+            $calendar->save();
 
-            for ($i=0; $i < count($arrTypeAdd); $i++) {
-                $time = new CalendarTime;
-                $time->type_ticket = $arrTypeAdd[$i];
-                $time->time_show = $arrTime[$i];
+            $arrayType = $request->types;
+            $arrayTypeID = $request->types_id;
+            $arrayTime = $request->time_shows;
+
+            for ($i=0; $i < count($arrayTypeID); $i++) {
+                $time = CalendarTime::find($arrayTypeID[$i]);
+                $time->type_ticket = $arrayType[$i];
+                $time->time_show = $arrayTime[$i];
                 $time->calendar_id = $calendar->id;
                 $time->save();
             }
-        }
 
-        // delete time if user choose
-        if ($request->has('calendarTimes_id')) {
-            foreach ($request->calendarTimes_id as $calTimeID) {
-                $calTime = CalendarTime::find($calTimeID);
-                $calTime->delete();
+            if ($request->has('types_add')) {
+                $arrTypeAdd = $request->types_add;
+                $arrTime = $request->time_shows_add;
+
+                for ($i=0; $i < count($arrTypeAdd); $i++) {
+                    $time = new CalendarTime;
+                    $time->type_ticket = $arrTypeAdd[$i];
+                    $time->time_show = $arrTime[$i];
+                    $time->calendar_id = $calendar->id;
+                    $time->save();
+                }
             }
+
+            // delete time if user choose
+            if ($request->has('calendarTimes_id')) {
+                foreach ($request->calendarTimes_id as $calTimeID) {
+                    $calTime = CalendarTime::find($calTimeID);
+                    $calTime->delete();
+                }
+            }
+            return back()->with('success', 'Sửa thành công');
+        } else {
+            return back()->withErrors('Lỗi, phim đã có lịch chiếu ngày '.$dateShow .' ở ' . $roomName->name . ' rạp ' . $theaterName[0]->name);
         }
-
-
-        return redirect('admin/calendars/'.$id.'/edit')->with('success', 'Sửa thành công');
     }
 
     /**
@@ -175,6 +204,13 @@ class CalendarController extends Controller
     {
         $calendar = Calendar::find($id);
         $calendar->calendarTimes()->delete();
+
+        foreach ($calendar->tickets as $ticket) {
+            foreach ($ticket->seats as $seat) {
+                $seat->delete();
+            }
+            $ticket->delete();
+        }
         $calendar->delete();
 
         return redirect('admin/calendars/')->with('success', 'Xóa thành công');
@@ -193,7 +229,8 @@ class CalendarController extends Controller
     public function ajaxShow($theaterID)
     {
         $theater = Theater::find($theaterID);
-        $calendars = $theater->calendars;
+        $calendars = $theater->calendars()->orderBy('date_show', 'asc')->get();
+
 
         $arrCalTimes = [];
         foreach ($calendars as $cal) {
@@ -201,7 +238,6 @@ class CalendarController extends Controller
                 $arrCalTimes[$cal->date_show][] = $calTimes;
             }
         }
-        //dd($arrCalTimes);
         return response()->json([
             'theater' => $theater,
             'caltimes' => $arrCalTimes,
